% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_util.R
\name{.appendList}
\alias{.appendList}
\title{Helper that appends a list item to a list of lists}
\usage{
.appendList(li = list(), item = list())
}
\arguments{
\item{li}{The container list}

\item{item}{The list to append to li as a new element}
}
\value{
The modified list
}
\description{
Helper that appends a list item to a list of lists
}
\keyword{internal}
