% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_delivery.R
\name{.countFilesInProduct}
\alias{.countFilesInProduct}
\title{Given a runId, polls the "download" method to count the number of files available
Uses HTTP HEAD to avoid downloading the files}
\usage{
.countFilesInProduct(self, runId = 0)
}
\arguments{
\item{self}{Caling object}

\item{runId}{Run id returned by .runDataProduct()}
}
\value{
(numeric) Number of files available for download
}
\description{
Given a runId, polls the "download" method to count the number of files available
Uses HTTP HEAD to avoid downloading the files
}
\keyword{internal}
