% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_discovery.R
\name{.discoveryRequest}
\alias{.discoveryRequest}
\title{Discovery Request
encapsulates a request to a discovery service}
\usage{
.discoveryRequest(self, filters = list(), service = "", method = "get")
}
\arguments{
\item{self}{Calling object}

\item{filters}{Named list of filters}

\item{service}{One of {"locations", "deployments", "deviceCategories", "devices",
"properties", "dataProducts"}}

\item{method}{One of {"get", "getTree"}}
}
\value{
(named list) Parsed response obtained
}
\description{
Discovery Request
encapsulates a request to a discovery service
}
\keyword{internal}
