% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_service.R
\name{.doRequest}
\alias{.doRequest}
\title{Generic request wrapper for making simple web service requests
Will return parsed results, even if they are an error description sent by the API
Will stop() if the request fails into an error different from 400}
\usage{
.doRequest(
  self,
  url = "",
  filters = list(),
  getInfo = FALSE,
  rawResponse = FALSE
)
}
\arguments{
\item{self}{Calling object}

\item{url}{Full URL to make the request, without GET parameters}

\item{filters}{GET parameters}

\item{getInfo}{When TRUE, will return a list with ("response", "duration", "status"), where
"duration" is the time the request took and "status" is the HTTP response status}

\item{rawResponse}{When TRUE, the response returned is the unparsed object returned
by httr::GET, otherwise a parsed named list is returned}
}
\value{
(list) The parsed response
}
\description{
Generic request wrapper for making simple web service requests
Will return parsed results, even if they are an error description sent by the API
Will stop() if the request fails into an error different from 400
}
\keyword{internal}
