% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_delivery.R
\name{.downloadDataProduct}
\alias{.downloadDataProduct}
\title{Public wrapper that lets a user download data products manually with a runId
Can optionally return just the download links}
\usage{
.downloadDataProduct(
  self,
  runId = 0,
  maxRetries = 0,
  downloadResultsOnly = FALSE,
  includeMetadataFile = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{self}{Calling object}

\item{runId}{(numeric) Run ID as provided by runDataProduct()}

\item{maxRetries}{(numeric) Maximum number of API requests allowed, 0 for no limit}

\item{downloadResultsOnly}{(logical) When TRUE, files are not downloaded
By default (FALSE) generated files are downloaded}

\item{includeMetadataFile}{(logical) When TRUE, a metadata file is downloaded,
otherwise it is skipped}

\item{overwrite}{(logical) When TRUE downloaded files will overwrite any file
with the same filename, otherwise they will be skipped}
}
\value{
(list) A list of results (one named list for each file) with
information on the operation outcome
}
\description{
Public wrapper that lets a user download data products manually with a runId
Can optionally return just the download links
}
\keyword{internal}
