% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_delivery.R
\name{.downloadProductFiles}
\alias{.downloadProductFiles}
\title{Download all data product files for provided run id}
\usage{
.downloadProductFiles(
  self,
  runId = 0,
  getMetadata = TRUE,
  maxRetries = 0,
  overwrite = TRUE,
  fileCount = 0
)
}
\arguments{
\item{self}{Caling object}

\item{runId}{Run id returned by .runDataProduct()}

\item{getMetadata}{When TRUE, the metadata file will be downloaded}

\item{maxRetries}{(numeric) Maximum number of API requests allowed, 0 for no limit}

\item{overwrite}{(logical) When TRUE downloaded files will overwrite any file
with the same filename, otherwise they will be skipped}

\item{fileCount}{The number of files to download, or 0 if unknown}
}
\value{
a list of named lists, with the results for each file
}
\description{
Download all data product files for provided run id
}
\keyword{internal}
