% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_delivery.R
\name{.estimatePollPeriod}
\alias{.estimatePollPeriod}
\title{Sets a poll period adequate to the estimated processing time
Longer processing times require longer poll periods to avoid going over maxRetries}
\usage{
.estimatePollPeriod(self, response)
}
\arguments{
\item{self}{Calling object}

\item{response}{Response obtained in .requestDataProduct() for the DP request}
}
\value{
(numeric) suggested time between server polls (seconds)
}
\description{
Sets a poll period adequate to the estimated processing time
Longer processing times require longer poll periods to avoid going over maxRetries
}
\keyword{internal}
