% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_archive.R
\name{.filterByExtension}
\alias{.filterByExtension}
\title{Filter file list results to only those where the filename ends with the extension
If extension is empty no change will be made}
\usage{
.filterByExtension(self, results = list(), extension = "")
}
\arguments{
\item{self}{Calling object}

\item{results}{Results as otained by getListByLocation() or getListByDevice()}

\item{extension}{(character) Extension to search for (i.e. "txt")}
}
\value{
Filtered list
}
\description{
Filter file list results to only those where the filename ends with the extension
If extension is empty no change will be made
}
\keyword{internal}
