% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_util.R
\name{.filters2url}
\alias{.filters2url}
\title{Returns a string URL for a base url and a named list of parameters}
\usage{
.filters2url(url, params)
}
\arguments{
\item{url}{Base url string}

\item{params}{Named list of parameters}
}
\value{
Full URL
}
\description{
Returns a string URL for a base url and a named list of parameters
}
\keyword{internal}
