% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_util.R
\name{.formatDuration}
\alias{.formatDuration}
\title{Returns a formatted time duration string representation of a duration in seconds}
\usage{
.formatDuration(secs = 0)
}
\arguments{
\item{secs}{(double) Number of seconds}
}
\value{
(character) A readable string for this time duration
}
\description{
Returns a formatted time duration string representation of a duration in seconds
}
\keyword{internal}
