% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_util.R
\name{.formatSize}
\alias{.formatSize}
\title{Returns a formatted file size string representation}
\usage{
.formatSize(size = 0)
}
\arguments{
\item{size}{(float) Size in bytes}
}
\value{
(character) a readable string for this file size
}
\description{
Returns a formatted file size string representation
}
\keyword{internal}
