% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_util.R
\name{.formatUtc}
\alias{.formatUtc}
\title{Format date as ISO8601 UTC}
\usage{
.formatUtc(self, dateString = "now")
}
\arguments{
\item{self}{Calling object}

\item{dateString}{A string that describes a date. Can also be "now"}
}
\value{
(character) Date string
}
\description{
Helper that returns an ISO8601 string for the provided date string
Most date formats are supported, as explained in: https://github.com/eddelbuettel/anytime
A value of "now" returns the current UTC date & time
Depends on the local system clock
}
\keyword{internal}
