% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_discovery.R
\name{.getDataProducts}
\alias{.getDataProducts}
\title{Request a list of data products that matches the filters provided}
\usage{
.getDataProducts(self, filters = list())
}
\arguments{
\item{self}{Calling object}

\item{filters}{(named list) search filters}
}
\value{
\preformatted{   (named list) A list of data products found, or an error description if any
}
}
\description{
Request a list of data products that matches the filters provided
}
\keyword{internal}
