% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_discovery.R
\name{.getDeviceCategories}
\alias{.getDeviceCategories}
\title{Request a list of device categories that matches the filters provided}
\usage{
.getDeviceCategories(self, filters = list())
}
\arguments{
\item{self}{Calling object}

\item{filters}{(named list) search filters}
}
\value{
\preformatted{   (named list) A list of device categories found, or an error description if any
}
}
\description{
Request a list of device categories that matches the filters provided
}
\keyword{internal}
