% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_realtime.R
\name{.getDirectAllPages}
\alias{.getDirectAllPages}
\title{Generic method to download and concatenate all pages of data
Keeps downloading all scalar or raw data pages until finished
Automatically translates sensorCategoryCodes to a string if a list is provided}
\usage{
.getDirectAllPages(
  self,
  filters = list(),
  service = "",
  method = "",
  allPages = FALSE
)
}
\arguments{
\item{self}{Calling object}

\item{filters}{(named list) Filters that describe the data origin}

\item{service}{One of: "scalardata", "rawdata"}

\item{method}{One of: "getByDevice", "getByLocation"}

\item{allPages}{When TRUE, if the data requested is too large to fit a single
API resquest, keep downloading data pages until we gather all data}
}
\value{
A single response in the expected format, with all data pages concatenated
}
\description{
Generic method to download and concatenate all pages of data
Keeps downloading all scalar or raw data pages until finished
Automatically translates sensorCategoryCodes to a string if a list is provided
}
\keyword{internal}
