% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_realtime.R
\name{.getDirectByDevice}
\alias{.getDirectByDevice}
\title{Obtains scalar data from a device, as described by the filters}
\usage{
.getDirectByDevice(self, filters = list(), allPages = FALSE)
}
\arguments{
\item{self}{Calling object}

\item{filters}{(named list) Filters that describe the data origin}

\item{allPages}{When TRUE, if the data requested is too large to fit a single
API resquest, keep downloading data pages until we gather all data}
}
\value{
(named list) Scalar data obtained for all sensors found
}
\description{
Obtains scalar data from a device, as described by the filters
}
\keyword{internal}
