% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_archive.R
\name{.getDirectFiles}
\alias{.getDirectFiles}
\title{Downloads all archive files that match the filters
Internally will use geListByDevice or getListByLocation and getFile all files}
\usage{
.getDirectFiles(self, filters = list(), overwrite = FALSE, allPages = FALSE)
}
\arguments{
\item{self}{Calling object}

\item{filters}{(named list) describe the data origin}

\item{overwrite}{When TRUE, downloaded files will overwrite any file with the
same filename, otherwise file will be skipped}

\item{allPages}{When TRUE, if the data is too long to fit a single request,
multiple pages will be requested until all data is obatined}
}
\value{
(named list) Information on the results of the operation, with "downloadResults"
for each file downloaded and general "stats"
}
\description{
Downloads all archive files that match the filters
Internally will use geListByDevice or getListByLocation and getFile all files
}
\keyword{internal}
