% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_archive.R
\name{.getDownloadUrl}
\alias{.getDownloadUrl}
\title{Given a filename, returns an archivefile absolute download URL}
\usage{
.getDownloadUrl(self, filename = "")
}
\arguments{
\item{self}{Calling object}

\item{filename}{(character) archive file name}
}
\value{
(character) download URL
}
\description{
Given a filename, returns an archivefile absolute download URL
}
\keyword{internal}
