% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_archive.R
\name{.getFile}
\alias{.getFile}
\title{Download the archive file with filename}
\usage{
.getFile(self, filename = "", overwrite = FALSE)
}
\arguments{
\item{self}{Calling object}

\item{filename}{Archive file filename}

\item{overwrite}{When TRUE, downloaded files will overwrite any file with the
same filename, otherwise file will be skipped}
}
\value{
(named list) Information on the download result
}
\description{
Download the archive file with filename
}
\keyword{internal}
