% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_archive.R
\name{.getList}
\alias{.getList}
\title{A generic wrapper for getListByLocation() and getListByDevice()}
\usage{
.getList(self, filters = list(), by = "location", allPages = FALSE)
}
\arguments{
\item{self}{Calling object}

\item{filters}{(named list) describe the data origin}

\item{by}{One of: "location", "device"}

\item{allPages}{When TRUE, if the data is too long to fit a single request,
multiple pages will be requested until all data is obatined}
}
\value{
(named list) Information on the list of files obtained
}
\description{
A generic wrapper for getListByLocation() and getListByDevice()
}
\keyword{internal}
