% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_archive.R
\name{.getListByLocation}
\alias{.getListByLocation}
\title{Get a list of files for a given location code and device category code, and
filtered by others optional parameters.}
\usage{
.getListByLocation(self, filters = list(), allPages = FALSE)
}
\arguments{
\item{self}{Calling object}

\item{filters}{(named list) describe the data origin}

\item{allPages}{When TRUE, if the data is too long to fit a single request,
multiple pages will be requested until all data is obatined}
}
\value{
(named list) file list obtained
}
\description{
Get a list of files for a given location code and device category code, and
filtered by others optional parameters.
}
\keyword{internal}
