% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_discovery.R
\name{.getLocationHierarchy}
\alias{.getLocationHierarchy}
\title{Request a hierarchical tree list of locations that matches the filters provided}
\usage{
.getLocationHierarchy(self, filters = list())
}
\arguments{
\item{self}{Calling object}

\item{filters}{(named list) search filters}
}
\value{
\preformatted{   (named list) A list of locations found, or an error description if any
           Locations include lists of children locations recursively
}
}
\description{
Request a hierarchical tree list of locations that matches the filters provided
}
\keyword{internal}
