% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_discovery.R
\name{.getLocations}
\alias{.getLocations}
\title{Request a list of locations that matches the filters provided}
\usage{
.getLocations(self, filters = list())
}
\arguments{
\item{self}{Calling object}

\item{filters}{(named list) search filters}
}
\value{
\preformatted{   (named list) A list of locations found, or an error description if any
}
}
\description{
Request a list of locations that matches the filters provided
}
\keyword{internal}
