% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_discovery.R
\name{.getProperties}
\alias{.getProperties}
\title{Request a list of properties that matches the filters provided}
\usage{
.getProperties(self, filters = list())
}
\arguments{
\item{self}{Calling object}

\item{filters}{(named list) search filters}
}
\value{
\preformatted{   (named list) A list of properties found, or an error description if any
}
}
\description{
Request a list of properties that matches the filters provided
}
\keyword{internal}
