% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_delivery.R
\name{.infoForProductFiles}
\alias{.infoForProductFiles}
\title{Returns a list of information lists for each file available for download
Returned rows will have the same structure as those returned by DataProductFile$getInfo()}
\usage{
.infoForProductFiles(self, dpRunId = 0, fileCount = 0, getMetadata = FALSE)
}
\arguments{
\item{self}{Caling object}

\item{dpRunId}{Run id returned by .runDataProduct()}

\item{fileCount}{The number of files to download, or 0 if unknown}

\item{getMetadata}{When TRUE, the metadata file will be included}
}
\value{
a list of named lists, with the results for each file
}
\description{
Returns a list of information lists for each file available for download
Returned rows will have the same structure as those returned by DataProductFile$getInfo()
}
\keyword{internal}
