% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_util.R
\name{.lastItem}
\alias{.lastItem}
\title{R helper: Returns the last element of a list}
\usage{
.lastItem(collection)
}
\arguments{
\item{collection}{The list}
}
\value{
The last element of the list
}
\description{
R helper: Returns the last element of a list
}
\keyword{internal}
