% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_util.R
\name{.log}
\alias{.log}
\title{Prints message to console only when showInfo is true}
\usage{
.log(self, msg = "")
}
\arguments{
\item{self}{Calling object}

\item{msg}{(character) The message to print}
}
\description{
Prints message to console only when showInfo is true
}
\keyword{internal}
