% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_delivery.R
\name{.orderDataProduct}
\alias{.orderDataProduct}
\title{Request, run and download a data product as described by the filters}
\usage{
.orderDataProduct(
  self,
  filters = list(),
  maxRetries = 0,
  downloadResultsOnly = FALSE,
  metadata = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{self}{Caling object}

\item{filters}{(named list) Filters that describe this data product}

\item{maxRetries}{(numeric) Total maximum number of request calls allowed, 0 for no limit}

\item{downloadResultsOnly}{(logical) When TRUE, files are not downloaded
By default (FALSE) generated files are downloaded}

\item{metadata}{(logical) When TRUE, a metadata file is downloaded,
otherwise it is skipped}

\item{overwrite}{(logical) When TRUE downloaded files will overwrite any file
with the same filename, otherwise they will be skipped}
}
\value{
(list) A list of results (one named list for each file) with
information on the operation outcome
}
\description{
Request, run and download a data product as described by the filters
}
\keyword{internal}
