% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_util.R
\name{.prepareDirectory}
\alias{.prepareDirectory}
\title{Creates directory if it does not exist
Supports directory paths with or without filenames}
\usage{
.prepareDirectory(filepath = "")
}
\arguments{
\item{filepath}{(character) Directory path}
}
\value{
(logical) TRUE if the directory was created or already exists
}
\description{
Creates directory if it does not exist
Supports directory paths with or without filenames
}
\keyword{internal}
