% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_util.R
\name{.prettyPrint}
\alias{.prettyPrint}
\title{Pretty prints a complex hierarchy of lists}
\usage{
.prettyPrint(item, name = "", level = 0, comma = FALSE, file = "")
}
\arguments{
\item{item}{Hierarchy item (can be a list or an atomic element)}

\item{name}{(character) Item name if its an element in a named list}

\item{level}{(numeric)   Depth level in the hierarchy, starting at 0}

\item{comma}{(logical)   If TRUE, a comma is printed after this element}

\item{file}{A file connection to an open file if we should print to it,
or "" (no file, print to console)}
}
\description{
Pretty prints a complex hierarchy of lists
}
\keyword{internal}
