% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_util.R
\name{.print}
\alias{.print}
\title{Helper for printing a JSON named list to the console
Can alternatively print the output to a text file at filePath}
\usage{
.print(self, obj, filePath = "")
}
\arguments{
\item{self}{self}

\item{obj}{named list to print (usually the result of another function)}

\item{filePath}{(string) if present, creates the file and writes the output in it}
}
\description{
Helper for printing a JSON named list to the console
Can alternatively print the output to a text file at filePath
}
\keyword{internal}
