% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_util.R
\name{.printErrorMessage}
\alias{.printErrorMessage}
\title{Print Error Response Message}
\usage{
.printErrorMessage(response)
}
\arguments{
\item{response}{Parsed response returned by httr::GET}
}
\description{
Prints the information from a API error response to the console
as a formatted error message
}
\keyword{internal}
