% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_delivery.R
\name{.printProductOrderStats}
\alias{.printProductOrderStats}
\title{Prints a formatted representation of the total time and size downloaded
after the product order finishes}
\usage{
.printProductOrderStats(self, fileList = list(), runInfo = list())
}
\arguments{
\item{self}{Calling object}

\item{fileList}{(list) As returned by .downloadProductFiles()}

\item{runInfo}{(list) As returned by .runDataProduct()}
}
\description{
Prints a formatted representation of the total time and size downloaded
after the product order finishes
}
\keyword{internal}
