% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_delivery.R
\name{.printProductRequest}
\alias{.printProductRequest}
\title{Prints the response after a data product request
The request response format might differ depending on the product origin
as it can be "assembled" on the fly, or reused from existing products}
\usage{
.printProductRequest(self, response)
}
\arguments{
\item{self}{Calling object}

\item{response}{Parsed httr response}
}
\description{
Prints the response after a data product request
The request response format might differ depending on the product origin
as it can be "assembled" on the fly, or reused from existing products
}
\keyword{internal}
