% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_delivery.R
\name{.requestDataProduct}
\alias{.requestDataProduct}
\title{Request a data product generation described by the filters}
\usage{
.requestDataProduct(self, filters = list())
}
\arguments{
\item{self}{Calling object}

\item{filters}{(named list) Filters that describe this data product}
}
\value{
Parsed httr response
}
\description{
Request a data product generation described by the filters
}
\keyword{internal}
