% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_util.R
\name{.respFailed}
\alias{.respFailed}
\title{Response has failed}
\usage{
.respFailed(response)
}
\arguments{
\item{response}{(named list) Parsed httr response}
}
\value{
(logical)
}
\description{
Returns TRUE if the response is a 401 JSON error description provided by the API
}
\keyword{internal}
