% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_util.R
\name{.saveAsFile}
\alias{.saveAsFile}
\title{Saves the file downloaded in the response object, in the outPath, with filename}
\usage{
.saveAsFile(response, filePath = "", fileName = "", overwrite = FALSE)
}
\arguments{
\item{response}{An http raw response as returned by httr::GET}

\item{filePath}{Path where the file will be saved}

\item{fileName}{Name of the file to save}

\item{overwrite}{If TRUE will overwrite files with the same name}
}
\value{
(numeric) Result code from {0: done, -1: error, -2: fileExists}
}
\description{
Saves the file downloaded in the response object, in the outPath, with filename
}
\keyword{internal}
