% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onc_service.R
\name{.serviceUrl}
\alias{.serviceUrl}
\title{Returns the absolute url for a given ONC API service}
\usage{
.serviceUrl(self, service = "")
}
\arguments{
\item{self}{Calling object}

\item{service}{One of ONC's API services}
}
\value{
(character) Service's URL
}
\description{
Returns the absolute url for a given ONC API service
}
\keyword{internal}
