% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipage.R
\docType{class}
\name{MultiPage-class}
\alias{MultiPage-class}
\alias{MultiPage}
\title{MultiPage}
\arguments{
\item{extension:}{Only provide for archivefiles filtering by filename extension}

\item{nextResponse}{(named list) Response with the next data page to add}

\item{responseTime:}{(double)     Request duration in seconds}

\item{response}{(named list) Response with the first page}

\item{service}{(character)  One of: "scalardata", "rawdata", "archivefiles"}
}
\value{
: Response with concatenated data for all pages obtained
Do a page request

Wraps the _doRequest method
Performs additional processing of the response for certain services

(list) A named list with: "response" (httr response) and "duration" (secs)
Concatenate data

Concatenates the data results from nextResponse into response
Compatible with the row structure of scalardata, rowdata and archivefiles

(named list) Modified original response
Estimate pages

Estimates the number of pages this request will require to download,
from the first page's response and its duration

(int) Estimated number of pages
Response Time Span

Determines the timespan covered by the data in the response

\preformatted{    (int)        page duration in seconds
}
}
\description{
API multiple page helper. Handles the download of data in multiple pages
Used by real-time and archivefile services
}
\section{Fields}{

\describe{
\item{\code{showInfo}}{As provided by the Onc class}

\item{\code{timeout}}{As provided by the Onc class
Class initializer
Get all pages

Obtains all data pages for a query with the filters
Multiple pages will be downloaded in sequence until completed
Each new page gets concatenated to the previous one (limited to RAM)}
}}


\keyword{internal}
